requirements.txt
Flask==3.0.2
deepface==0.0.83
Werkzeug==3.0.2
numpy==1.23.5
tensorflow==2.10.0
keras==2.10.0
opencv-python==4.9.0.80
matplotlib==3.8.4
pillow==10.3.0
